requirements.txt
streamlit>=1.36
opencv-python-headless>=4.8
numpy>=1.24
Pillow>=10.0
matplotlib>=3.7
